@@ -3,25 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "apexcharts": "^3.26.0",
-    "axios": "^0.21.1",
-    "formik": "^1.5.1",
-    "google-map-react": "^2.1.9",
-    "js-base64": "^3.6.0",
-    "moment": "^2.24.0",
-    "react": "^16.8.4",
-    "react-apexcharts": "^1.3.7",
-    "react-datepicker": "^3.6.0",
-    "react-dom": "^16.8.4",
-    "react-icons": "^4.2.0",
-    "react-loader-spinner": "^4.0.0",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.8",
-    "react-select": "^4.3.0",
-    "react-social-login": "^3.4.13",
-    "react-toastify": "^7.0.3"
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.1",
+    "@mui/material": "^5.15.1",
+    "apexcharts": "^3.45.0",
+    "axios": "^1.6.2",
+    "formik": "^2.4.5",
+    "google-map-react": "^2.2.1",
+    "js-base64": "^3.7.5",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-apexcharts": "^1.4.1",
+    "react-datepicker": "^4.25.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.12.0",
+    "react-loader-spinner": "^6.1.5",
+    "react-router-dom": "^6.21.1",
+    "react-scripts": "^5.0.1",
+    "react-select": "^5.8.0",
+    "react-social-login": "^3.4.17",
+    "react-toastify": "^9.1.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -46,6 +48,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {}
+  }
 }
