@@ -1,67 +1,103 @@
-import React, { Component } from 'react'
-import SidebarRowComponent from './SidebarRowComponent'
-import HomeWorkOutlinedIcon from '@material-ui/icons/HomeWorkOutlined';
-import ContactMailIcon from '@material-ui/icons/ContactMail';
-import BarChartIcon from '@material-ui/icons/BarChart';
-import { NavLink } from 'react-router-dom'
-import SettingsIcon from '@material-ui/icons/Settings';
-import AccountCircleIcon from '@material-ui/icons/AccountCircle';
-import AuthenticationService from "../api/AuthenticationService"
-import SecurityIcon from '@material-ui/icons/Security';
-import {FaInfoCircle} from 'react-icons/fa'
-import DirectionsWalkIcon from '@material-ui/icons/DirectionsWalk';
-import ArrowBackIcon from '@material-ui/icons/ArrowBack';
-import { GENERAL_INFO_LINK_URL, CHANGE_PASSWORD_LINK_URL } from '../Constants'
-import LocationCityIcon from '@material-ui/icons/LocationCity';
-class SidebarComponent extends Component {
+import React, { useState, useEffect } from 'react';
+import SidebarRowComponent from './SidebarRowComponent';
+import HomeWorkOutlinedIcon from '@mui/icons-material/HomeWorkOutlined';
+import ContactMailIcon from '@mui/icons-material/ContactMail';
+import BarChartIcon from '@mui/icons-material/BarChart';
+import { NavLink, useLocation } from 'react-router-dom';
+import SettingsIcon from '@mui/icons-material/Settings';
+import AccountCircleIcon from '@mui/icons-material/AccountCircle';
+import AuthenticationService from '../api/AuthenticationService';
+import SecurityIcon from '@mui/icons-material/Security';
+import { FaInfoCircle } from 'react-icons/fa';
+import DirectionsWalkIcon from '@mui/icons-material/DirectionsWalk';
+import ArrowBackIcon from '@mui/icons-material/ArrowBack';
+import { GENERAL_INFO_LINK_URL, CHANGE_PASSWORD_LINK_URL } from '../Constants';
+import LocationCityIcon from '@mui/icons-material/LocationCity';
 
-    constructor(props) {
-        super(props)
-        this.state = {
-            settingsMenu: false
-        }
-    }
+const SidebarComponent = () => {
+  const location = useLocation();
+  const [settingsMenu, setSettingsMenu] = useState(false);
 
-    componentDidMount() {
-        this.updateStateOfSidebar();
-    }
+  useEffect(() => {
+    updateStateOfSidebar();
+  }, [location.pathname]);
 
-    componentDidUpdate() {
-        this.updateStateOfSidebar();
+  const updateStateOfSidebar = () => {
+    if (
+      location.pathname !== GENERAL_INFO_LINK_URL &&
+      location.pathname !== CHANGE_PASSWORD_LINK_URL &&
+      settingsMenu
+    ) {
+      setSettingsMenu(false);
+    } else if (
+      (location.pathname === GENERAL_INFO_LINK_URL && !settingsMenu) ||
+      (location.pathname === CHANGE_PASSWORD_LINK_URL && !settingsMenu)
+    ) {
+      setSettingsMenu(true);
     }
+  };
 
-    updateStateOfSidebar() {
-        if (window.location.pathname !== GENERAL_INFO_LINK_URL && window.location.pathname !== CHANGE_PASSWORD_LINK_URL
-            && this.state.settingsMenu) {
-            this.setState({ settingsMenu: false })
-        }
-        else if ((window.location.pathname === GENERAL_INFO_LINK_URL && !this.state.settingsMenu)
-            || (window.location.pathname === CHANGE_PASSWORD_LINK_URL && !this.state.settingsMenu)) {
-            this.setState({ settingsMenu: true });
-        }
-    }
+  const onClick = () => {
+    setSettingsMenu(true);
+  };
 
-    onClick() {
-        this.setState({ settingsMenu: true });
-    }
+  const isLoggedIn = AuthenticationService.isUserLoggedIn();
 
-    render() {
-        const isLoggedIn = AuthenticationService.isUserLoggedIn();
-        return (
-            <div className="sidebar">
-                {!this.state.settingsMenu && <NavLink to="/" exact activeClassName="active"><SidebarRowComponent icon={BarChartIcon} title="Dashboard"></SidebarRowComponent> </NavLink>}
-                {isLoggedIn && !this.state.settingsMenu && <NavLink to="/home" activeClassName="active" onClick={this.onClick.bind(this)}><SidebarRowComponent icon={HomeWorkOutlinedIcon} title="Home Data" /></NavLink>}
-                {isLoggedIn && !this.state.settingsMenu && <NavLink to="/location" activeClassName="active" onClick={this.onClick.bind(this)}><SidebarRowComponent icon={DirectionsWalkIcon} title="Walks Data" /></NavLink>}
-                {!this.state.settingsMenu && <NavLink to="/city" exact activeClassName="active"><SidebarRowComponent icon={LocationCityIcon} title="City Data"></SidebarRowComponent> </NavLink>}
-                {isLoggedIn && !this.state.settingsMenu && <NavLink to={GENERAL_INFO_LINK_URL} activeClassName="active" onClick={this.onClick.bind(this)}><SidebarRowComponent icon={SettingsIcon} title="Settings" /></NavLink>}
-                {!this.state.settingsMenu && <NavLink to="/info" exact activeClassName="active"><SidebarRowComponent icon={FaInfoCircle} title="Info"></SidebarRowComponent> </NavLink>}
-                {!this.state.settingsMenu && <NavLink to="/contact" activeClassName="active" onClick={this.onClick.bind(this)}><SidebarRowComponent icon={ContactMailIcon} title="Contact" /></NavLink>}
-                
-                {this.state.settingsMenu && <NavLink to={GENERAL_INFO_LINK_URL} activeClassName="active" onClick={this.onClick.bind(this)}><SidebarRowComponent icon={AccountCircleIcon} title="General" /></NavLink>}
-                {this.state.settingsMenu && <NavLink to={CHANGE_PASSWORD_LINK_URL} activeClassName="active" onClick={this.onClick.bind(this)}><SidebarRowComponent icon={SecurityIcon} title="Security" /></NavLink>}
-                {this.state.settingsMenu && <NavLink to="/" exact activeClassName="active" onClick={this.onClick.bind(this)}><SidebarRowComponent icon={ArrowBackIcon} title="Back" /></NavLink>}
-            </div>
-        )
-    }
-}
+  return (
+    <div className="sidebar">
+      {!settingsMenu && (
+        <NavLink to="/" exact activeClassName="active">
+          <SidebarRowComponent icon={BarChartIcon} title="Dashboard" />
+        </NavLink>
+      )}
+      {isLoggedIn && !settingsMenu && (
+        <NavLink to="/home" activeClassName="active" onClick={onClick}>
+          <SidebarRowComponent icon={HomeWorkOutlinedIcon} title="Home Data" />
+        </NavLink>
+      )}
+      {isLoggedIn && !settingsMenu && (
+        <NavLink to="/location" activeClassName="active" onClick={onClick}>
+          <SidebarRowComponent icon={DirectionsWalkIcon} title="Walks Data" />
+        </NavLink>
+      )}
+      {!settingsMenu && (
+        <NavLink to="/city" exact activeClassName="active">
+          <SidebarRowComponent icon={LocationCityIcon} title="City Data" />
+        </NavLink>
+      )}
+      {isLoggedIn && !settingsMenu && (
+        <NavLink to={GENERAL_INFO_LINK_URL} activeClassName="active" onClick={onClick}>
+          <SidebarRowComponent icon={SettingsIcon} title="Settings" />
+        </NavLink>
+      )}
+      {!settingsMenu && (
+        <NavLink to="/info" exact activeClassName="active">
+          <SidebarRowComponent icon={FaInfoCircle} title="Info" />
+        </NavLink>
+      )}
+      {!settingsMenu && (
+        <NavLink to="/contact" activeClassName="active" onClick={onClick}>
+          <SidebarRowComponent icon={ContactMailIcon} title="Contact" />
+        </NavLink>
+      )}
+
+      {settingsMenu && (
+        <NavLink to={GENERAL_INFO_LINK_URL} activeClassName="active" onClick={onClick}>
+          <SidebarRowComponent icon={AccountCircleIcon} title="General" />
+        </NavLink>
+      )}
+      {settingsMenu && (
+        <NavLink to={CHANGE_PASSWORD_LINK_URL} activeClassName="active" onClick={onClick}>
+          <SidebarRowComponent icon={SecurityIcon} title="Security" />
+        </NavLink>
+      )}
+      {settingsMenu && (
+        <NavLink to="/" exact activeClassName="active" onClick={onClick}>
+          <SidebarRowComponent icon={ArrowBackIcon} title="Back" />
+        </NavLink>
+      )}
+    </div>
+  );
+};
+
 export default SidebarComponent;
